+++ src/components/LeftBar.vue
@@ -33,6 +33,7 @@ const emit = defineEmits<{
   saveFileAs: [];
   exportPdf: [];
   exportDocx: [];
+  presentMarp: [];
   toggleCodeView: [];
   toggleSplit: [];
   toggleSplitEditor: [];
@@ -74,6 +75,7 @@ const emit = defineEmits<{
         @save-file-as="emit('saveFileAs')"
         @export-pdf="emit('exportPdf')"
         @export-docx="emit('exportDocx')"
+        @present-marp="emit('presentMarp')"
         @toggle-code-view="emit('toggleCodeView')"
         @toggle-split="emit('toggleSplit')"
         @toggle-split-editor="emit('toggleSplitEditor')"