+++ src/components/StatusBar.vue
@@ -30,6 +30,7 @@ const emit = defineEmits<{
   saveFileAs: [];
   exportPdf: [];
   exportDocx: [];
+  presentMarp: [];
   toggleCodeView: [];
   toggleSplit: [];
   toggleSplitEditor: [];
@@ -66,6 +67,7 @@ const emit = defineEmits<{
         @save-file-as="emit('saveFileAs')"
         @export-pdf="emit('exportPdf')"
         @export-docx="emit('exportDocx')"
+        @present-marp="emit('presentMarp')"
         @toggle-code-view="emit('toggleCodeView')"
         @toggle-split="emit('toggleSplit')"
         @toggle-split-editor="emit('toggleSplitEditor')"